requirement.txt
certifi==2019.3.9
chardet==3.0.4
et-xmlfile==1.0.1
idna==2.8
jdcal==1.4.1
lxml==4.3.3
numpy==1.16.3
openpyxl==2.6.2
pandas==0.24.2
Pillow==6.0.0
python-dateutil==2.8.0
python-docx==0.8.10
pytz==2019.1
requests==2.22.0
six==1.12.0
urllib3==1.25.2
xlrd==1.2.0
